requirements.txt
openpyxl~=3.1.5
python-telegram-bot~=21.7
selenium~=4.26.1
pandas~=2.2.3